All the actual Matlab file of the simulator
Step #1: Simulate the trajectory with a lot of environment variable take in consideration. These calculations will be repeated in loop by the IA to found the perfect trajectory to hit the target. MATLAB was choosing because it's very easy to debug and very easy to program the calculation in 4D.